- made size a size_t in map2cs AddFile. - added %ignore in cspace.i for csStringBase::operator[]. - changed strnicmp in csShaderGLCGVP::Compile to csStrNCaseCmp.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20641 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/import/map2cs/zipfile.cpp
@@ -35,7 +35,7 @@ bool CZipFile::AddFile(const char* originalname, const char* zipname)
 
   //Check the size of the file.
   fseek(fd, 0, SEEK_END); //Seek to end of file
-  int size = ftell(fd);   //Check current position
+  size_t size = ftell(fd);   //Check current position
   fseek(fd, 0, SEEK_SET); //Seek to beginning of file
 
   void* file = NewFile(zipname, size);
